* master: (462 commits)
  ofxThreadedImageLoader: fix problem with new ofThread impl
  ofxThreadedImageLoader: fixes for std::thread
  ofThread: fix issue with merged branch
  remove orig file from repo
  ofAppendUTF8: uint32 parameter instead of int
  ofVec2f: fix warning
  Git ignore folders and contents with apps prefix in root only
  round again
  check subfolder to get back to
  back to check
  fix ignore for subfolders
  ofEasyCam: Remove unneeded comment
  ofEasyCam: Fix tabs
  ofEasyCam: update comments to break at <= 80 chars
  Add rotation and translation sensitivity setters
  add any folder with name prefix apps to gitignore
  release script overwrite folder if exists
  create_package: fix makefile copying for linux arm
  Do not rely on a specific version of OpenCv
  Do not rely on a specific version of libwebp
  ...

# Conflicts:
#	addons/ofxAndroid/src/ofxAndroidVideoGrabber.cpp

+++ scripts/linux/cleanAllExamples.sh
@@ -3,25 +3,54 @@
 for category in $( find ../../examples/ -maxdepth 1 -type d )
 do
     if [ "$category" = "../../examples/android" -o "$category" = "../../examples/ios" -o "$category" = "../../examples/" ]; then
-            continue
+        continue
     fi
-    
-    for example in $( find $category -maxdepth 1 -type d | grep -v osx )
+
+    for example in $( find "$category" -maxdepth 1 -type d | grep -v osx )
     do
-	    if [ "$example" = "$category" ]; then
-		    continue
-	    fi
-        echo "-----------------------------------------------------------------"
-        echo "cleaning " + $example
-        
-	    make clean -C $example
-        rm -rf $example/obj 2> /dev/null
-        rm -rf $example/*.layout 2> /dev/null
-        rm -rf $example/*.backup 2> /dev/null
-        rm -rf $example/*.depend 2> /dev/null
-        rm $example/*~ 2> /dev/null
-        
-        echo "-----------------------------------------------------------------"
-        echo ""
+        if [ "$example" = "$category" ]; then
+            continue
+        fi
+        echo -----------------------------------------------------------------
+        echo cleaning "$example"
+
+        make clean -C "$example"
+        rm -rf "${example:?}/obj" 2> /dev/null
+        rm -rf "${example:?}/"*.layout 2> /dev/null
+        rm -rf "${example:?}/"*.backup 2> /dev/null
+        rm -rf "${example:?}/"*.depend 2> /dev/null
+        rm "${example:?}/"*~ 2> /dev/null
+	if [ -f ${example}.qbs ]; then
+	    qbs clean
+	fi
+	rm "{example:?}/"*.qbs 2> /dev/null
+        echo -----------------------------------------------------------------
+        echo
+    done
+done
+
+for category in $( find ../../apps/ -maxdepth 1 -type d )
+do
+    for example in $( find "$category" -maxdepth 1 -type d | grep -v osx )
+    do
+        if [ "$example" = "$category" ]; then
+            continue
+        fi
+        echo -----------------------------------------------------------------
+        echo cleaning "$example"
+
+        make clean -C "$example"
+        rm -rf "${example:?}/obj" 2> /dev/null
+        rm -rf "${example:?}/"*.layout 2> /dev/null
+        rm -rf "${example:?}/"*.backup 2> /dev/null
+        rm -rf "${example:?}/"*.depend 2> /dev/null
+        rm "${example:?}/"*~ 2> /dev/null
+        if [ -f ${example}.qbs ]; then
+            qbs clean
+        fi
+        rm "{example:?}/"*.qbs 2> /dev/null
+
+        echo -----------------------------------------------------------------
+        echo
     done
 done
